Refactor `count` to `count_documents` and `estimated_document_count`
PyMongo `~=4.0` removes `Cursor.count` and `Collection.count` and provides
two new count methods: `Collection.count_documents` and
`Collection.estimated_document_count`.

Both `Cursor.count` and `Collection.count` in PyMongo `~=3.0` use `count` MongoDB
command, which provide potentially inaccurate results when executd without a query
predicate, in a transaction, or on a sharded cluster.

`Collection.count_documents` issues an aggregate query, which is accurate in
all cases, but relatively expensive.

`Collection.estimated_document_count` in PyMongo `>=4.0,<4.2` uses a $collStats
aggregation to estimate the document count for the collection. In PyMongo `>4.2`
it relies on the `count` MongoDB command.

When possible, we now try using the estimated document count for count operations,
but if a limit or skip is defined, or a hint is provided, we'll use the more accurate
count.

This mirrors the behavior of the upstream mongoengine implementation.